update para spaceship com delta times

+++ Projecto_v1/Proj_v1/Alien.cpp
@@ -16,7 +16,7 @@ Alien::~Alien() {
 
 }
 
-void Alien::update(float delta) {
+void Alien::update(int delta) {
 
 }
